CH/keys refactored
