+++ bin/www
@@ -7,13 +7,14 @@
 var app = require('../src/app');
 var debug = require('debug')('kodehauz-training-portal:server');
 var http = require('http');
-const chalk = require('chalk')
+const chalk = require('chalk');
+const keys = require('./../config/keys');
 
 /**
  * Get port from environment and store in Express.
  */
 
-var port = normalizePort(process.env.PORT);
+var port = normalizePort(keys.PORT);
 app.set('port', port);
 
 /**
